Rename ALL THE THINGS! \o/
