+++ .eslintrc.json
@@ -23,6 +23,7 @@
     "import/no-extraneous-dependencies": ["error", { "devDependencies": ["**/*.test.*", "./*"] }],
     "jsx-a11y/label-has-associated-control": ["error", { "assert": "either" }],
     "no-shadow": "off",
+    "no-unused-vars": "off",
     "react/jsx-filename-extension": ["error", { "extensions": [".tsx"] }],
     "react/no-unknown-property": [
       "error",
