+++ src/components/daily-auths-report.tsx
@@ -1,76 +1,82 @@
 import { VNode } from "preact";
 import { useContext, useRef, useState } from "preact/hooks";
-import { utcDays, utcDay } from "d3-time";
+import { utcDay } from "d3-time";
 import * as Plot from "@observablehq/plot";
-import { format } from "d3-format";
 import { useQuery } from "preact-fetching";
-import { utcFormat } from "d3-time-format";
-import { group, ascending, rollup } from "d3-array";
 import Markdown from "preact-markdown";
-import { ReportFilterContext } from "./context/report-filter-context";
+import { ascending, group, rollup } from "d3-array";
+import { ReportFilterContext } from "../contexts/report-filter-context";
 import Table, { TableData } from "./table";
-import { path as reportPath } from "./report";
 import PlotComponent from "./plot";
-import { useAgencies } from "./context/agencies-context";
+import { useAgencies } from "../contexts/agencies-context";
 import Accordion from "./accordion";
-import useResizeListener from "./hooks/resize-listener";
+import useResizeListener from "../hooks/resize-listener";
+import { ProcessedResult, loadData } from "../models/daily-auths-report-data";
+import { formatSIDropTrailingZeroes, formatWithCommas, yearMonthDayFormat } from "../formats";
 
-interface Result {
-  count: number;
+function plot({
+  start,
+  finish,
+  data,
+  agency,
+  ial,
+  facetAgency,
+  width,
+}: {
+  start: Date;
+  finish: Date;
+  data?: ProcessedResult[];
+  agency?: string;
+  ial: number;
+  facetAgency?: boolean;
+  width?: number;
+}): HTMLElement {
+  return Plot.plot({
+    height: facetAgency ? new Set((data || []).map((d) => d.agency)).size * 60 : undefined,
+    width,
+    y: {
+      tickFormat: formatSIDropTrailingZeroes,
+    },
+    x: {
+      domain: [start, finish],
+    },
+    facet: facetAgency
+      ? {
+          data: data || [],
+          y: "agency",
+          marginRight: 150,
+        }
+      : undefined,
+    style: {},
+    marks: [
+      Plot.ruleY([0]),
+      Plot.rectY(
+        data || [],
+        Plot.binX(
+          {
+            y: "sum",
+          },
+          {
+            y: "count",
+            x: {
+              value: "date",
+              thresholds: utcDay,
+            },
+            fill: agency ? "friendly_name" : "steelblue",
+            title: (bin: ProcessedResult[]) => {
+              const date = yearMonthDayFormat(bin[0].date);
+              const total = formatWithCommas(bin.reduce((sum, d) => sum + d.count, 0));
+              const friendlyName = bin[0]?.friendly_name;
 
-  ial: 1 | 2;
-
-  issuer: string;
-
-  /**
-   * This is always present but we don't use it, so easier to mark as optional
-   */
-  iaa?: string;
-
-  // eslint-disable-next-line camelcase
-  friendly_name: string;
-
-  agency: string;
+              return [agency && `${friendlyName}:`, total, `(${date})`].filter(Boolean).join(" ");
+            },
+            filter: (d: ProcessedResult) => d.ial === ial && (!agency || d.agency === agency),
+          }
+        )
+      ),
+    ],
+  });
 }
-
-interface DailyAuthsReportData {
-  results: Result[];
-
-  /**
-   * ISO8601 string
-   */
-  start: string;
-
-  /**
-   * ISO8601 string
-   */
-  finish: string;
-}
-
-export interface ProcessedResult extends Result {
-  date: Date;
-}
-
-function process(report: DailyAuthsReportData): ProcessedResult[] {
-  const date = new Date(report.start);
-  return report.results.map((r) => ({ ...r, date, agency: r.agency || "(No Agency)" }));
-}
-
-function loadData(
-  start: Date,
-  finish: Date,
-  env: string,
-  fetch = window.fetch
-): Promise<ProcessedResult[]> {
-  return Promise.all(
-    utcDays(start, finish, 1).map((date) => {
-      const path = reportPath({ reportName: "daily-auths-report", date, env });
-      return fetch(path).then((response) => response.json());
-    })
-  ).then((reports) => reports.flatMap((r) => process(r)));
-}
-
-const yearMonthDayFormat = utcFormat("%Y-%m-%d");
 
 function tabulate({
   results,
@@ -181,74 +187,6 @@ function tabulateSumByAgency({
   };
 }
 
-const formatWithCommas = format(",");
-const formatSIPrefix = format(".2s");
-const formatSIDropTrailingZeroes = (d: number): string => formatSIPrefix(d).replace(/\.0+/, "");
-
-function plot({
-  start,
-  finish,
-  data,
-  agency,
-  ial,
-  facetAgency,
-  width,
-}: {
-  start: Date;
-  finish: Date;
-  data?: ProcessedResult[];
-  agency?: string;
-  ial: number;
-  facetAgency?: boolean;
-  width?: number;
-}): HTMLElement {
-  return Plot.plot({
-    height: facetAgency ? new Set((data || []).map((d) => d.agency)).size * 60 : undefined,
-    width,
-    y: {
-      tickFormat: formatSIDropTrailingZeroes,
-    },
-    x: {
-      domain: [start, finish],
-    },
-    facet: facetAgency
-      ? {
-          data: data || [],
-          y: "agency",
-          marginRight: 150,
-        }
-      : undefined,
-    style: {},
-    marks: [
-      Plot.ruleY([0]),
-      Plot.rectY(
-        data || [],
-        Plot.binX(
-          {
-            y: "sum",
-          },
-          {
-            y: "count",
-            x: {
-              value: "date",
-              thresholds: utcDay,
-            },
-            fill: agency ? "friendly_name" : "steelblue",
-            title: (bin: ProcessedResult[]) => {
-              const date = yearMonthDayFormat(bin[0].date);
-              const total = formatWithCommas(bin.reduce((sum, d) => sum + d.count, 0));
-              const friendlyName = bin[0]?.friendly_name;
-
-              return [agency && `${friendlyName}:`, total, `(${date})`].filter(Boolean).join(" ");
-            },
-            filter: (d: ProcessedResult) => d.ial === ial && (!agency || d.agency === agency),
-          }
-        )
-      ),
-    ],
-  });
-}
-
 function DailyAuthsReport(): VNode {
   const ref = useRef(null as HTMLDivElement | null);
   const [width, setWidth] = useState(undefined as number | undefined);
@@ -295,4 +233,4 @@ twice will count twice. It does not de-duplicate users or provide unique auths.`
 }
 
 export default DailyAuthsReport;
-export { tabulate, tabulateSumByAgency, loadData, formatSIDropTrailingZeroes };
+export { tabulate, tabulateSumByAgency };
