+++ src/components/plot.tsx
